requirements.txt
Baker==1.3
argparse==1.2.1
cachet==0.1.0
docker-py==1.8.0
requests==2.9.1
six==1.10.0
toml==0.9.1
websocket-client==0.37.0
